+++ src/Feliz.Listeners/CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 2.0.0-rc.2 - 2025-10-02
+
+### 🔄 Changed 
+
+- Updated for `FsReact` change in Feliz
+
 ## 2.0.0-rc.1 - 2025-10-02
 
 ### Changed