+++ src/Feliz/CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 3.0.0-rc.9 - 2025-11-03
+
+### 🔄 Changed
+
+- Moved `IDisposable` helper into `FsReact` type. This type will be home for any future custom hooks. by @Freymaurer
+
 ## 3.0.0-rc.8 - 2025-11-03
 
 ### 🐛 Fixed